restruct layout
